Add errors
These are errors that can be thrown by the interfaces in this
repository. Contributors to the project should be able to catch these
errors, so they need to be available at the base level.

The `NotImplementedError` and `AbstractInstantiationError` were both
originally written as part of the appliance-core project, but are being
moved here.

Issue #1 Move errors from appliance-core