Add Payload object
The Payload object is fully implemented because it represents an entity
that is used on the fringe of all components.  It is possible this
should become duplicated code instead of existing in the base
repository.

Payloads will be serialized to json and transmitted (e.g. via kafka or
via SocketIO).

Issue #3 Create Payload object