Add data type constants
The dataTypes.js constant file holds the list of standard data types
supported in the TV Kitchen ecosystem. It is possible for an appliance
developer to emit types of data outside of this list in their local
environment, but the network will only recognize data that fits within
one of these types.

This commit does not create an exhaustive list of data types, but simply
sets up the structure for future development.

Issue #7 Create a dataTypes constant file