Disable import/named plugin for now
This is dirty but I think it's our only choice.

The 'named' rule of the ESLint import plugin seems to have a bug which
doesn't properly detect named aggregate exports (e.g. `export * as
constants from ...`).

I opened an issue in the plugin project, which may eventually be
resolved (https://github.com/benmosher/eslint-plugin-import/issues/1753)

In the meantime we just need to disable the rule to move forward.