Move to monorepo architecture
This is a massive commit, but we'll see how it comes out in the wash...

It moves code to monorepos, while also adjusting the import syntax for
isolated projects (e.g. IAppliance imports the base-errors package
instead of using a relative import).

The lint and build commands are adjusted to reflect the new world order
(in the case of lint, it's just a better command that we should use
elsewhere).

There is also now a `yarn bootstrap` command which is important for
setting up the dependency links between sibling repositories.

Issue #9