Rename stream container data type
We realized that `STREAM.DATA` is a bit more generic than we need to be,
so we're getting *slightly* more specific by renaming that variable to
`STREAM.CONTAINER`.

There was some exploration of the concept of getting even more concrete
with something along the lines of `STREAM.CONTAINER.MPEGTS` but since we
are not planning to support other types of containers in the near future
this felt like it would be more likely wrong than right.

Issue #14 Replace STREAM.DATA

Co-authored-by: Justin Reese <justin@justinreese.com>