Add time attributes to Payload
We need to be able to keep track of time related to the data in
a Payload.

We wanted to be able to expose how far into a stream a payload is
associated with, as well as what moment of time associated with
a payload.

We opted to use milliseconds for duration-based attributes, and ISO 6081 for
timestamp based attributes.

Issue #34 Add time to Payload