Update IAppliance invoke method
The previous definition of invoke put the burden of looking up the
Payloads that it would process on the method itself. Similarly, it
required the `invoke` method to manage the internal state (either
directly or via method calls) to register Payloads that would need to be
re-processed.

By updating the `invoke` API, we are suggesting that that management
should occur outside of the `invoke` logic itself. It will be passed the
Payloads that it is being asked to process, and it will return the
Payloads that were not processed but still warrant future processing.

The change to parameters should not impact anybody who implemented using
the old API, though it is recommended that all future implementations of
`invoke` avoid managing the Appliance's internal Payload state directly.

The change to the return value is a breaking change, since the previous
definition returned a Boolean, not a PayloadArray.

Issue #42