Add logger to IAppliance
Appliances are going to need to be able to log things, especially as we
work towards the concept of a method that verifies if an Appliance has
dependencies properly installed and outputs resolution instructions.

This adds a logger with an API that will work with
a [winston](https://www.npmjs.com/package/winston)  object.

There are methods for each of the log levels specified in the TV Kitchen
base constant `logLevel`.

Appliance developers should not assume that logger has been replaced,
and should only call the methods that are officially supported in the
IAppliance documentation.

Issue #48