Add sortedArray utilities
We are going to support a sorted payload array data type, and that will
involve a few methods that benefit from searches that are optimized
around sorted arrays. These methods do not need to be invented, but we
also don't want to bring in bloated dependencies.

After looking at the method-level lodash dependencies we found that the
best path would be to implement our own version of sortedIndexBy. The
lodash dependency represented 2,000 lines of code, whereas our
implementation is under 20.