Add IAppliance duck typing
We can't rely on instanceof to detect implementations of IAppliance in
other parts of the project; this creates an alternative method to detect
instances of IAppliance.

Note that since we cannot instantiate IAppliance directly, we had to
create a helper "class" in order to use our getOwnPropertyNames short
cut.  The alternative was to allow a special flag such as
`allowInterfaceInstantiation` that could be passed in as an
overrideSetting, but I preferred the isolated code smell to something
that would actually impact the API of IAppliance.

Issue #57