Remove emitResult from IAppliance
`emitResult` was always documented as an internal helper method, but
since IAppliance did not actually use it, it did not belong at that
level.

IAppliance should really only be defining the external API -- for
anybody *using* an Appliance. `emitResult` was a method for an Appliance
implementor, which means it belongs in `AbstractAppliance` instead.

Issue #67