Convert type lookup methods to static
The getInputType and getOutputType methods of IAppliance were originally
defined as being instance methods.  In reality, we should not have to
instantiate an appliance in order to know what kinds of data its
instances will consume and produce -- those attributes are really
attributes of the appliance itself.

This becomes particularly important when implementing the countertop's
topology definition algorithms.

Issue #77