Update IAppliance to extend Transform
Appliances used to communicate new data by emitting payloads. The
problem with this model is that there is no concept of backpressure,
which means if an appliance was processing data slower than it was
receiving it (or if it was generating data faster than the countertop
could handle it), there was risks of memory leaks and eventually system
failure.

By extending Transform, appliances become pipe-able objects that can
leverage all of the built in backpressure logic in Node.

The `ingestPayload` method has been modified to reflect the transform
API, though it should never be called directly and instead the `.write`
method that comes from Transform is the new point of payload insertion
to an appliance.

Issue #88