Add Payload data helpers
The Payload object has a `data` attribute which must be a buffer.
However, there are many times when we want to associate more useful data
types within that buffer (such as a string or a simple JavaScript
object).

This adds new setter / getter helper methods to the Payload class in
order to more consistently populate and access a payload's data.

Issue #96