Replace timestamp with origin
Prior to this commit, Payload objects were expected to have
a `timestamp` field which represented the absolute time that the payload
was associated with.  This was very unwieldy to calculate, and forced
every appliance to internally keep track of an "origin timestamp" for
a given stream.

We're flipping that situation on its head, instead including the
payload's `origin` timestamp.  This change impacts AvroPayload and
PayloadArray as well.

Issue #102